% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstas.R
\name{dstats}
\alias{dstats}
\title{Create Histogram, Boxplot and Numeric Summary}
\usage{
dstats(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Create Histogram, Boxplot and Numeric Summary
}
