% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstats.R
\name{dstats}
\alias{dstats}
\title{Creates histogram, boxplot, line chart, density graph and descriptive summary}
\usage{
dstats(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Creates histogram, boxplot, line chart, density graph and descriptive summary
}
